🚚 Move `sparseArray` arbitrary into /arbitrary
